+++ land-queue/config.py
@@ -1,3 +1,5 @@
+import os
+
 WEBHOOKS_LOG_FILE = "logs/webhooks.json"
 KEY_FILE = "monorepoapptest.private-key.pem"
-APP_ID = "64638"
+APP_ID = os.environ.get("APP_ID", "64638")
